API (server) Skeleton - Section 2
Created a WebApi template using Microsoft Entity Framework 6.0. Added a database using SqLite and seeded the file with roughly 20 products. So far the server only has two methods to query the database. One to get all the products in the database, and another to get a single product by the product id.