React (client) Skeleton - Section 3
Created a React template to use for client side processing. I also imported Material UI and used that for styling the items from the database into cards on the screen. Added a navigation bar (AppBar) also using MUI with a switch on it to toggle the website from dark mode to light mode and vice versa. Basic functionality is now setup on both the client and server side.