React Router - Section 4
For this commit, I installed React Router and setup a routing page for the home, about, contact, catalog and product details web pages. I also styled the AppBar and added a few Nav Links so I could access the different pages by clicking on the links instead of typing in the web browser. Pretty simple section so the commit is very small.