Shopping Cart Feature - Section 6
Added a fully functional checkout feature on both the API (server) and React (client). You can now add items to the shopping cart from both the catalog and product detail pages for any quantity you wish to update. But in a temporary fill in page for the checkout page, but it won't be fully functional until I can get authorization working in a later commit.