Paging, Sorting & Filtering - Section 8
Adding sorting, search and filter functionality to both the API (server) and the client (React). Also added a pagination component to the bottom of the client page to see the results of filtering and searching. Very simple section to complete without too much boilerplate thanks to using states in Redux.