Added Identity - Section 9
Adding Identity allows the website to support logins from different users. Users must be logged in (and have at least 1 item in the shopping cart) to be able to go to the checkout page. Users can logout at anytime and upon login (within 7 days) the shopping cart will return with previous items. Anonymous users can start a shopping cart, but again, must login to be able to check those items out. Having said that, the next logical feature to work on is the checkout page.