Added Checkout Feature - Section
Finally added a checkout feature to the website that creates a unique order number ID and loads it with the content of the shopping cart. Once the order is created, the shopping cart is cleared and the website will then ask for payment for the order (which obviously doesn't work yet). Next commit will focus on giving the website the ability to accept payments.